Centralizing a few of the memory-related constants to avoid some redundancies and also to have a single place to update the memory map if needed.
See Task 3507.

Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/330
Tested-by: Jenkins Server

+++ src/kernel/heapmgr.C
@@ -102,7 +102,7 @@ void HeapManager::newPage()
 {
     void* page = PageManager::allocatePage();
     chunk_t * c = (chunk_t*)page;
-    for (int i = 0; i < (PAGESIZE / (1 << (BUCKETS + 3))); i++)
+    for (uint64_t i = 0; i < (PAGESIZE / (1 << (BUCKETS + 3))); i++)
     {
 	c->len = BUCKETS-1;
 	push_bucket(c, BUCKETS-1);
